requirements.txt
black
flake8
isort
loguru
mkdocs
pip
python-dotenv
tqdm
typer
torch
torchvision
-e .
